pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "sigmorph"
version = "0.1.0"
description = "Generate explainable Sigma rules from suspicious Windows process events."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "PyYAML>=6.0",
  "typer>=0.12"
]

[project.scripts]
sigmorph = "sigmorph.cli:app"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
